requirements.txt
opencv-python-headless
numpy
torch
torchvision
Pillow
matplotlib
pandas
PyYAML
requests
tqdm
psutil
py-cpuinfo
thop
seaborn
scikit-learn
ultralytics
filterpy
scipy
roboflow
inference_sdk
streamlit
python-dotenv
albumentations